backend/requirements.txt
# =============================================================
# Python Dependencies - Face Recognition Backend
# =============================================================
# Instalasi:
# 1. Buka terminal/PowerShell di folder backend
# 2. Jalankan: D:/xampp/htdocs/facereconigtion/.venv/Scripts/pip.exe install -r requirements.txt
# =============================================================

# Web Framework
Flask==3.1.2
Werkzeug==3.1.3

# Computer Vision & Face Recognition
opencv-python==4.12.0.88
face-recognition==1.3.0
face-recognition-models==0.3.0
Pillow==12.0.0

# dlib (diperlukan oleh face-recognition)
# Untuk Windows, gunakan pre-built binary:
dlib-bin==19.24.6

# MQTT Communication (untuk ESP32)
paho-mqtt==2.1.0

# HTTP Requests
requests==2.32.5

# Utilities
numpy==2.2.6
click==8.3.0
